qomplnt-api-1/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
alembic
PyJWT
python-dotenv
mysqlclient



